Decimal no float

+++ models/price_history_model.py
@@ -1,6 +1,7 @@
 from enum import Enum
 from datetime import datetime
 from pydantic import BaseModel, Field
+from decimal import Decimal
 
 
 class PriceChangeType(str, Enum):
@@ -12,6 +13,6 @@ class PriceHistory(BaseModel):
   id: str = Field(..., alias="_id")
   product_id: str = Field(...)
   type: PriceChangeType = PriceChangeType.PURCHASE
-  price: float = Field(...)
+  price: Decimal = Field(..., decimal_places=2)
   date: datetime = Field(...)
   created_at: datetime = None
