+++ models/purchase_detail_model.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel, Field
 from datetime import datetime
+from decimal import Decimal
 
 
 class Product(BaseModel):
@@ -13,7 +14,7 @@ class PurchaseDetail(BaseModel):
   purchase_id: str = Field(...)
   product: Product = Field(...)
   units: int = Field(...)
-  unity_price: float = Field(...)
-  total_price: float = Field(...)
+  unity_price: Decimal = Field(..., decimal_places=2)
+  total_price: Decimal = Field(..., decimal_places=2)
   created_at: datetime = None
   updated_at: datetime = None
