+++ schemas/product_schema.py
@@ -1,6 +1,7 @@
-from pydantic import BaseModel, Field
 from datetime import datetime
+from pydantic import BaseModel, Field
 from schemas.query_base import QueryBase
+from decimal import Decimal
 
 
 class ProductCreate(BaseModel):
@@ -23,8 +24,8 @@ class ProductCreateResponse(BaseModel):
   code: str = Field(..., max_length=8)
   description: str = Field(...)
   categories: list[str] = Field(...)
-  purchase_price: float = Field(...)
-  sale_price: float = Field(...)
+  purchase_price: Decimal = Field(..., decimal_places=2)
+  sale_price: Decimal = Field(..., decimal_places=2)
   stock: int = Field(default=0)
   created_at: datetime = None
   updated_at: datetime = None
