+++ schemas/sale_schema.py
@@ -1,16 +1,17 @@
 from typing import Union, Tuple
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel, Field, field_validator
 from schemas.query_base import QueryBase
+from schemas.utils import parse_amount_query
 from models.sale_model import Sale, PayWith
 from models.sale_detail_model import SaleDetail
-from schemas.utils import parse_amount_query
 
 
 class Product(BaseModel):
   id: str = Field(..., alias="_id")
   units: str = Field(...)
-  price: float = Field(...)
+  price: Decimal = Field(..., decimal_places=2)
 
 
 class SaleCreate(BaseModel):
